pyproject.toml
[tool.poetry]
name = "arc-finetuning-st"
version = "0.1.0"
description = ""
authors = ["Andrei Fajardo <andrei@llamaindex.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
llama-index-core = "^0.11.9"
llama-index-llms-openai = "^0.2.7"
llama-index-agent-openai = "^0.3.1"
llama-index-program-openai = "^0.2.0"
llama-index-readers-file = "^0.2.1"
llama-deploy = {version = "^0.1.3", extras = ["redis"]}
streamlit = "^1.38.0"
llama-index-llms-anthropic = "^0.3.1"


[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.2.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
